Only run Remove-OldInstances for Wmi option

+++ src/PowerShell/FETCH_no_logs.ps1
@@ -960,7 +960,7 @@ try {
         $jsonOutput | Out-File $WriteTo
     }   
 
-    if ($DeleteOlderThanDays -ne $null) {
+    if ($Wmi -and $DeleteOlderThanDays -ne $null) {
         Remove-OldInstances -DeleteOlderThanDays $DeleteOlderThanDays -WmiNamespace $WmiNamespace -WmiClassName "$WmiClassPrefix`Sessions"
         Remove-OldInstances -DeleteOlderThanDays $DeleteOlderThanDays -WmiNamespace $WmiNamespace -WmiClassName "$WmiClassPrefix`UserRights"
         Remove-OldInstances -DeleteOlderThanDays $DeleteOlderThanDays -WmiNamespace $WmiNamespace -WmiClassName "$WmiClassPrefix`LocalGroups"
